+++ server/src/weaverbird/pipeline/steps/join.py
@@ -38,7 +38,7 @@ class JoinStepWithRef(BaseJoinStep):
     ) -> JoinStepWithVariable | None:
         right_pipeline = await resolve_if_reference(reference_resolver, self.right_pipeline)
         if right_pipeline is None:
-            from ..references import ReferenceUnresolved
+            from weaverbird.pipeline.pipeline import ReferenceUnresolved
 
             raise ReferenceUnresolved()
         return JoinStepWithVariable(